+++ src/types.ts
@@ -48,6 +48,7 @@ export namespace types {
   export declare type everyFn = (node: treeNode, index: int) => boolean
   export declare type nodeToStringFn = (node: treeNode) => string
   export declare type formatFunction = (val: string, rowIndex: positiveInt, colIndex: positiveInt) => string
+  export declare type nodeIdRenameMap = { [oldNodeTypeId: string]: string }
 
   export declare type children = string | Object | treeNode | any // todo: specify better.